Automated/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
scipy>=1.7.0
prophet>=1.0
scikit-learn>=0.24.0
requests>=2.26.0
python-dotenv>=0.19.0 